advanced_api_project/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.5
drf-yasg==1.21.7
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.1
